спринт 1 2.4.1

+++ 01 Основные компоненты C++/01 Спринт 1/02 Базовые алгоритмы/04 Алгоритм accumulate/Задание 1/src/main.cpp
@@ -12,7 +12,7 @@ double Average(const vector<int>& xs) {
     if (size == 0) {
         return 0;
     }
-    return double(accumulate(xs.begin(), xs.end(), 0)) / size;
+    return accumulate(xs.begin(), xs.end(), 0.0) / size;
 }
 
 int main() {
